+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+#### Changed
+- Use consistent primary blue color for all notification types
+
 ## [5.0.3] - 2025-06-13
 
 #### Fixed